SIG-873 move generic API functionality
