+++ api/app/tests/apps/signals/test_filters.py
@@ -9,7 +9,7 @@ from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.test import APITestCase
 from rest_framework.viewsets import GenericViewSet
 
-from signals.apps.signals.filters import FieldMappingOrderingFilter
+from signals.apps.signals.api_generics.filters import FieldMappingOrderingFilter
 from signals.apps.signals.models import MainCategory, Priority, Signal, SubCategory
 from signals.apps.signals.v0.serializers import SignalAuthHALSerializer
 from tests.apps.signals.factories import SignalFactory
